feat: add client-side GitHub username validation
